moved render Peers to Video.js changed Test.js to Video.js commented out Peer JS Video part cause not working yet
