+++ frontend/src/components/App/App.js
@@ -15,7 +15,7 @@ import {
 import Login from '../Login/Login';
 import Home from '../Home/Home';
 import Room from '../Room/Room';
-import Test from '../Test/Test';
+import Video from '../Video/Video';
 
 
 // Import Socket API
